+++ src/common_cpp/DataStructure/V1731.cc
@@ -159,11 +159,11 @@ void V1731::SetMaxPos(int max_pos) {
     _max_pos = max_pos;
 }
 
-double V1731::GetTriggerTimeTag() const {
+int V1731::GetTriggerTimeTag() const {
     return _trigger_time_tag;
 }
 
-void V1731::SetTriggerTimeTag(double trigger_time_tag) {
+void V1731::SetTriggerTimeTag(int trigger_time_tag) {
     _trigger_time_tag = trigger_time_tag;
 }
 
@@ -199,11 +199,11 @@ void V1731::SetGeo(int geo) {
     _geo = geo;
 }
 
-int V1731::GetPedestal() const {
+double V1731::GetPedestal() const {
     return _pedestal;
 }
 
-void V1731::SetPedestal(int pedestal) {
+void V1731::SetPedestal(double pedestal) {
     _pedestal = pedestal;
 }
